Move all remaining wallet db representations
