+++ key-manager/src/main/scala/org/bitcoins/keymanager/bip39/BIP39KeyManager.scala
@@ -11,7 +11,7 @@ import org.bitcoins.core.api.keymanager.{
 import org.bitcoins.core.compat.{CompatEither, CompatLeft, CompatRight}
 import org.bitcoins.core.crypto._
 import org.bitcoins.core.hd.{HDAccount, HDPath}
-import org.bitcoins.core.util.{BitcoinSLogger, TimeUtil}
+import org.bitcoins.core.util.{BitcoinSLogger, HDUtil, TimeUtil}
 import org.bitcoins.core.wallet.keymanagement.KeyManagerUnlockError._
 import org.bitcoins.core.wallet.keymanagement.{
   InitializeKeyManagerError,
@@ -20,7 +20,6 @@ import org.bitcoins.core.wallet.keymanagement.{
 }
 import org.bitcoins.crypto.{AesPassword, Sign}
 import org.bitcoins.keymanager._
-import org.bitcoins.keymanager.util.HDUtil
 import scodec.bits.BitVector
 
 import scala.util.{Failure, Success, Try}
