+++ wallet-test/src/test/scala/org/bitcoins/wallet/EncryptedMnemonicTest.scala
@@ -3,9 +3,9 @@ package org.bitcoins.wallet
 import org.bitcoins.core.util.TimeUtil
 import org.bitcoins.crypto.AesPassword
 import org.bitcoins.keymanager.{DecryptedMnemonic, EncryptedMnemonicHelper}
+import org.bitcoins.testkit.Implicits._
 import org.bitcoins.testkit.core.gen.CryptoGenerators
 import org.bitcoins.testkit.util.BitcoinSUnitTest
-import org.bitcoins.testkit.Implicits._
 
 import scala.util.{Failure, Success}
