+++ wallet/src/main/scala/org/bitcoins/wallet/internal/TransactionProcessing.scala
@@ -1,11 +1,6 @@
 package org.bitcoins.wallet.internal
 
-import org.bitcoins.core.api.wallet.db.{
-  AddressTagDb,
-  SpendingInfoDb,
-  TransactionDb,
-  TransactionDbHelper
-}
+import org.bitcoins.core.api.wallet.db._
 import org.bitcoins.core.currency.CurrencyUnit
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.BitcoinAddress
@@ -17,7 +12,6 @@ import org.bitcoins.core.wallet.utxo.{AddressTag, TxoState}
 import org.bitcoins.crypto.{DoubleSha256Digest, DoubleSha256DigestBE}
 import org.bitcoins.wallet._
 import org.bitcoins.wallet.api.{AddUtxoError, AddUtxoSuccess}
-import org.bitcoins.wallet.models._
 
 import scala.concurrent.{Future, Promise}
 import scala.util.{Failure, Success, Try}
