+++ wallet/src/main/scala/org/bitcoins/wallet/internal/UtxoHandling.scala
@@ -1,11 +1,6 @@
 package org.bitcoins.wallet.internal
 
-import org.bitcoins.core.api.wallet.db.{
-  LegacySpendingInfo,
-  NestedSegwitV0SpendingInfo,
-  SegwitV0SpendingInfo,
-  SpendingInfoDb
-}
+import org.bitcoins.core.api.wallet.db._
 import org.bitcoins.core.compat._
 import org.bitcoins.core.hd.HDAccount
 import org.bitcoins.core.number.UInt32
@@ -24,7 +19,6 @@ import org.bitcoins.core.util.{EitherUtil, FutureUtil}
 import org.bitcoins.core.wallet.utxo._
 import org.bitcoins.crypto.DoubleSha256DigestBE
 import org.bitcoins.wallet.api.{AddUtxoError, AddUtxoResult, AddUtxoSuccess}
-import org.bitcoins.wallet.models._
 import org.bitcoins.wallet.{Wallet, WalletLogger}
 
 import scala.concurrent.Future
