+++ wallet/src/main/scala/org/bitcoins/wallet/models/IncomingTransactionDAO.scala
@@ -1,6 +1,6 @@
 package org.bitcoins.wallet.models
 
-import org.bitcoins.core.api.wallet.db.TransactionDb
+import org.bitcoins.core.api.wallet.db.{IncomingTransactionDb, TransactionDb}
 import org.bitcoins.core.currency.CurrencyUnit
 import org.bitcoins.crypto.DoubleSha256DigestBE
 import org.bitcoins.wallet.config._
